// src/types/partials/CodeableConcept.ts
// This file is auto-generated by generate_types so do not edit manually

import { TExtension } from '../partials/Extension';
import { TCoding } from '../partials/Coding';

export type TCodeableConcept = {
    id?: String;
    extension?: TExtension[];
    coding?: TCoding[];
    text?: String;
};